change type of the edit_type

+++ src/main/resources/db/initDB_postgre.sql
@@ -170,7 +170,7 @@ CREATE TABLE week_days
  */
 CREATE TABLE edit_types
 (
-  id   SERIAL PRIMARY KEY,
+  id   SMALLINT PRIMARY KEY,
   code VARCHAR(255) UNIQUE NOT NULL
 );
